Use IProjectThreadingService in NuGetUIThreadHelper
Update the NuGetUIThreadHelper to use IProjectThreadingService for VS 2017 and IThreadHandling for VS 2015. These JoinableTaskFactories are able to work with the CPS project system used by .NETCore projects and avoid hangs.

Fixes https://github.com/NuGet/Home/issues/4185
