+++ src/NuGet.Clients/NuGet.Tools/NuGetPackage.cs
@@ -158,6 +158,9 @@ namespace NuGetVSExtension
             _dteEvents = _dte.Events.DTEEvents;
             _dteEvents.OnBeginShutdown += OnBeginShutDown;
 
+            // Set the JoinableTaskFactory for the current version of visual studio.
+            NuGetUIThreadHelper.SetJoinableTaskFactoryFromService(ServiceLocator.PackageServiceProvider);
+
             SetDefaultCredentialProvider();
 
             if (SolutionManager.NuGetProjectContext == null)
