+++ src/NuGet.Clients/VisualStudio.Facade/project.json
@@ -1,6 +1,7 @@
 ﻿{
   "dependencies": {
-    "Microsoft.VisualStudio.Shell.Interop": "7.10.6071"
+    "Microsoft.VisualStudio.Shell.Interop": "7.10.6071",
+    "Microsoft.VisualStudio.Shell.Interop.11.0": "11.0.61030"
   },
   "frameworks": {
     "net46": {}
